requirements.txt
streamlit==1.28.0
pandas==2.0.3
plotly==5.17.0
textblob==0.17.1
wordcloud==1.9.2
matplotlib==3.7.2
apify-client==1.4.0
python-dotenv==1.0.0
requests==2.31.0
# # ──────────────────────────────────────────────────────────────
# # Multi-Source Review Sentiment Analyzer – Requirements
# # ──────────────────────────────────────────────────────────────
# # Install: pip install -r requirements.txt

# # ── Web Framework ──────────────────────────────────────────────
# fastapi==0.111.0
# uvicorn[standard]==0.30.1
# python-multipart==0.0.9      # for file uploads in FastAPI

# # ── Database / ORM ────────────────────────────────────────────
# sqlalchemy==2.0.30

# # ── NLP – VADER ───────────────────────────────────────────────
# vaderSentiment==3.3.2

# # ── NLP – BERT / Transformers ─────────────────────────────────
# transformers==4.41.2
# torch>=2.6.0                  # CPU build – swap for GPU if available
# # torch==2.3.0+cu118 --index-url https://download.pytorch.org/whl/cu118

# # ── Scraping ───────────────────────────────────────────────────
# requests==2.32.3
# beautifulsoup4==4.12.3
# lxml==5.2.2                   # fast HTML parser for BS4
# selenium==4.21.0              # optional – needs chromedriver in PATH
# webdriver-manager==4.0.1      # auto-manages chromedriver

# # ── Frontend ───────────────────────────────────────────────────
# streamlit==1.35.0
# plotly==5.22.0
# pandas==2.2.2

# # ── Language Detection & Translation ─────────────────────────
# langdetect==1.0.9
# deep-translator==1.11.4

# # ── Utilities ──────────────────────────────────────────────────
# python-dotenv==1.0.1
# pydantic==2.7.3
# httpx==0.27.0
